requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
flask-cors==6.0.0
PyJWT==2.8.0
Werkzeug==3.1.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.41
